Refactor component imports and class names 🧰
